move cel_event_type module
